+++ src/main/java/com/yoProgramo/BackEnd/repository/ProvinciaRepository.java
@@ -1,12 +1,12 @@
 
 package com.yoProgramo.BackEnd.repository;
 
-import com.yoProgramo.BackEnd.entity.Provincia;
+import com.yoProgramo.BackEnd.model.Provincia;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ProvinciaRepository extends JpaRepository<Provincia, Integer>{
+public interface ProvinciaRepository extends JpaRepository<Provincia, Long>{
     
     
 }
